Moved Function Components into their own files
