+++ src/App.js
@@ -1,220 +1,51 @@
 import { useState } from "react";
 import eventsData from "./data";
-//import { v1 as generateUniqueID } from "uuid";
-//import Attendees from "./Attendees";
-//import Event from "./Components/Event";
-import Footer from "./Components/Footer";
 import Header from "./Components/Header";
-//import NewEventForm from "./Components/NewEventForm";
+import Event from "./Components/Event";
+import NewEventForm from "./Components/NewEventForm";
+import Footer from "./Components/Footer";
 
-function App() {
-  //const [events, setEvents] = useState(eventsData); // moved to Events.js file
 
-  //const [showAttendees, setShowAttendees] = useState(false); // moved to Attendees.js file
-
-  const [selectOption, setSelectOption] = useState("");
-
-  // const [newEvent, setNewEvent] = useState({
-  //   id: "",
-  //   eventType: "",
-  //   name: "",
-  //   organizer: "",
-  //   eventImage: "",
-  //   date: "",
-  //   people: [],
-  // });   // moved to NewEventsForm.js file
-
-  // function addEvent() { 
-  //   const createEvent = {
-  //     id: generateUniqueID(),
-  //     eventType: selectOption,
-  //     name: newEvent.name,
-  //     organizer: newEvent.organizer,
-  //     eventImage: newEvent.eventImage || "https://loremflickr.com/640/480/",
-  //     date: newEvent.date,
-  //     people: [],
-  //   };
-  //   handleAddEvent(createEvent);
-  // }      // moved to NewEventForm.js file
-
-  function handleSelectChange(e) {
-    setSelectOption(e.target.value);
-  }
-
-  // function handleSubmit(e) {
-  //   e.preventDefault();
-  //   addEvent();
-  //   resetEventForm();
-  // }     // moved to Event.js file
-
-  function handleTextChange(e) {
-    setNewEvent({
-      ...newEvent,
-      [e.target.id]: e.target.value,
-    });
-  }
-
-  // function resetEventForm() {
-  //   setNewEvent({
-  //     id: "",
-  //     eventType: "",
-  //     name: "",
-  //     organizer: "",
-  //     eventImage: "",
-  //     date: "",
-  //   });
-  //   setSelectOption("");
-  // }     // moved to NewEventForm.js file
+  function App() {
+    const [events, setEvents] = useState(eventsData);
 
   function handleAddEvent(event) {
     setEvents([event, ...events]);
   }
 
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
+  function updateEventAttendance(eventId, attendeeId) {
+    const eventArray = [...events];
+    const eventIndex = eventArray.findIndex((event) => eventId === event.id);
+    const event = { ...eventArray[eventIndex] };
+    const personIndex = event.people.findIndex(
+      (person) => person.id === attendeeId
+    );
+    const peopleArray = [...event.people];
+    peopleArray[personIndex].attendance = !peopleArray[personIndex].attendance;
+    event.people = peopleArray;
+    eventArray[eventIndex] = event;
+    setEvents(eventArray);
   }
-
-  // function updateEventAttendance(eventId, attendeeId) { // This file was moved to Attendee.js file
-  //   const eventArray = [...events];
-  //   const eventIndex = eventArray.findIndex((event) => eventId === event.id);
-  //   const event = { ...eventArray[eventIndex] };
-  //   const personIndex = event.people.findIndex(
-  //     (person) => person.id === attendeeId
-  //   );
-  //   const peopleArray = [...event.people];
-  //   peopleArray[personIndex].attendance = !peopleArray[personIndex].attendance;
-  //   event.people = peopleArray;
-  //   eventArray[eventIndex] = event;
-  //   setEvents(eventArray);
-  // }
 
   return (
     <div className="App">
-      <>
-        <header>
-          <h1 className="color-change-5x">RSVP App</h1>
-        </header>
-      </>
+      <Header />
       <main>
-        <div className="new-event">
-          <>
-            <form onSubmit={handleSubmit}>
-              <h3>Create a new event</h3>
-              <label htmlFor="name">Event name:</label>
-              <input
-                type="text"
-                id="name"
-                onChange={handleTextChange}
-                value={newEvent.name}
-              />
-
-              <label htmlFor="organizer">Organizer:</label>
-              <input
-                type="text"
-                id="organizer"
-                onChange={handleTextChange}
-                value={newEvent.organizer}
-              />
-
-              <label htmlFor="eventImage">Event image:</label>
-              <input
-                type="text"
-                id="eventImage"
-                onChange={handleTextChange}
-                value={newEvent.eventImage}
-              />
-              <label htmlFor="eventType">Event type:</label>
-              <select id="eventType" onChange={handleSelectChange}>
-                <option value=""></option>
-                <option value="Birthday">Birthday</option>
-                <option value="Anniversary">Anniversary</option>
-                <option value="Intramural Sport">Intramural Sport</option>
-                <option value="Watch Party">Watch Party</option>
-                <option value="wedding">Wedding</option>
-              </select>
-              <br />
-              <input type="submit" />
-            </form>
-          </>
+        <div className="new-event">  
+          <NewEventForm handleAddEvent={handleAddEvent} />
         </div>
         <div className="events">
           <ul>
-            {events.map((event) => {
-              const { people: attendees } = event;
-
-              return (
-                <>
-                  <li key={event.id}>
-                    <img src={event.eventImage} alt={event.name} />
-                    <h5>
-                      {event.name} {event.eventType}
-                    </h5>
-                    <br />
-                    <span>Organized by: {event.organizer} </span>
-                    <br />
-                    <>
-                      <button onClick={toggleEventAttendees}>
-                        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
-                      </button>
-
-                      {showAttendees ? (
-                        <div className="attendees">
-                          {attendees.map((attendee, index) => (
-                            <>
-                              <div key={attendee.id} className="attendee">
-                                <p>
-                                  <img
-                                    src={attendee.avatar}
-                                    alt={attendee.firstName}
-                                  />
-                                  {"   "}
-                                  <span>
-                                    {" "}
-                                    {attendee.firstName} {attendee.lastName}{" "}
-                                  </span>
-                                </p>
-                                <p>
-                                  <button
-                                    className="clickable"
-                                    onClick={() =>
-                                      updateEventAttendance(
-                                        event.id,
-                                        attendee.id
-                                      )
-                                    }
-                                  >
-                                    Attending:
-                                  </button>
-                                  <span>
-                                    {attendee.attendance ? "✅" : "❌"}
-                                  </span>
-                                </p>
-
-                                <p>
-                                  <span>Note:</span> {attendee.note}
-                                </p>
-                              </div>
-                            </>
-                          ))}
-                        </div>
-                      ) : null}
-                    </>
-                  </li>
-                </>
-              );
-            })}
+          {events.map((event) => (
+              <Event
+                event={event}
+                updateEventAttendance={updateEventAttendance}
+              />
+            ))}
           </ul>
         </div>
       </main>
-      <>
-        <footer>
-          <ul>
-            <li>Contact</li>
-            <li>About</li>
-            <li>Legal</li>
-          </ul>
-        </footer>
-      </>
+      <Footer />
     </div>
   );
 }
