+++ src/Components/Attendees.js
@@ -1,5 +1,31 @@
 import { useState } from "react";
-export default function Attendees() {
+import Attendee from "./Attendee";
+
+export default function Attendees({ event, updateEventAttendance }) {
   const [showAttendees, setShowAttendees] = useState(false);
-  return;
+
+  const { people: attendees } = event;
+
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  }
+  return (
+    <>
+      <button onClick={toggleEventAttendees}>
+        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+      </button>
+
+      {showAttendees ? (
+        <div className="attendees">
+          {attendees.map((attendee, index) => (
+            <Attendee
+              attendee={attendee}
+              event={event}
+              updateEventAttendance={updateEventAttendance}
+            />
+          ))}
+        </div>
+      ) : null}
+    </>
+  );
 }
