+++ src/Components/NewEventForm.js
@@ -1,5 +1,9 @@
 import { useState } from "react";
+import { v1 as generateUniqueID } from "uuid";
+
 export default function NewEventForm({ handleAddEvent }) {
+  const [selectOption, setSelectOption] = useState("");
+
   const [newEvent, setNewEvent] = useState({
     id: "",
     eventType: "",
@@ -9,7 +13,7 @@ export default function NewEventForm({ handleAddEvent }) {
     date: "",
     people: [],
   });
-  
+
   function addEvent() { 
     const createEvent = {
       id: generateUniqueID(),
@@ -23,6 +27,23 @@ export default function NewEventForm({ handleAddEvent }) {
     handleAddEvent(createEvent);
   }
 
+  function handleSelectChange(e) {
+    setSelectOption(e.target.value);
+  }
+
+  function handleSubmit(e) {
+    e.preventDefault();
+    addEvent();
+    resetEventForm();
+  }
+
+  function handleTextChange(e) {
+    setNewEvent({
+      ...newEvent,
+      [e.target.id]: e.target.value,
+    });
+  }
+
   function resetEventForm() {
     setNewEvent({
       id: "",
@@ -34,5 +55,46 @@ export default function NewEventForm({ handleAddEvent }) {
     });
     setSelectOption("");
   }
-  return;
+
+  return (
+    <>
+    <form onSubmit={handleSubmit}>
+      <h3>Create a new event</h3>
+      <label htmlFor="name">Event name:</label>
+      <input
+        type="text"
+        id="name"
+        onChange={handleTextChange}
+        value={newEvent.name}
+      />
+
+    <label htmlFor="organizer">Organizer:</label>
+    <input
+      type="text"
+      id="organizer"
+      onChange={handleTextChange}
+      value={newEvent.organizer}
+    />
+
+<label htmlFor="eventImage">Event image:</label>
+        <input
+          type="text"
+          id="eventImage"
+          onChange={handleTextChange}
+          value={newEvent.eventImage}
+        />
+        <label htmlFor="eventType">Event type:</label>
+        <select id="eventType" onChange={handleSelectChange}>
+          <option value=""></option>
+          <option value="Birthday">Birthday</option>
+          <option value="Anniversary">Anniversary</option>
+          <option value="Intramural Sport">Intramural Sport</option>
+          <option value="Watch Party">Watch Party</option>
+          <option value="wedding">Wedding</option>
+        </select>
+        <br />
+        <input type="submit" />
+      </form>
+    </>
+  );
 }
